Refactoring code, adding missing markdown changes
