+++ src/data/events/talks/blockchain-101.md
@@ -1,0 +1,24 @@
+---
+title: Blockchain 101 - Decentralization, Cryptoeconomics, and ICOs
+speakers:
+  - name: Júlio Santos
+    occupations:
+      - what: CTO
+        where: Fractal
+
+      - what: Co-Founder
+        where: Life on Mars
+    bio: Júlio is a software engineer who started 4 companies. He's one of the founders of Life on Mars, a Porto-based software team with sky-high standards and aspirations. He's also the CTO of Fractal, a Berlin-based blockchain tech company. To this day, he has no idea what he's doing, but somehow keeps merrily chugging along in blissful oblivion.
+
+    img: "../../../images/gatsby-icon.png"
+    linkedin: https://www.linkedin.com/in/jvsantos/
+
+day: 2019-11-01
+start_time: 17h10 
+end_time: 17h50
+place: B032
+path: /blockchain-101
+type: 0-talk
+---
+
+Blockchain technology, coupled with cryptoeconomics, allows us to confidently approach a whole new class of problems. It's a force for good, bringing censorship resistance to a global economy. In this short talk, I'll give you a short overview of the ecosystem, covering decentralization, Bitcoin, Ethereum, and ICOs
