+++ src/data/events/talks/test1.md
@@ -1,0 +1,34 @@
+---
+title: Software Engineering at Outsystems
+
+speakers:
+  - name: Ricardo Araújo
+    occupations:
+      - what: senior software engineer
+        where: outsystems
+    bio: Almadense de gema, Proencense de coração, Ricardo Araújo tem 20 anos de carreira ligado ao desenvolvimento web, tendo aprendido e desenvolvido aplicações para diversas indústrias e clientes de diversos países. Em 2009 abriu o centro de desenvolvimento remoto da OutSystems situado em Proença-a-Nova e de lá tem ajudado a OutSystems a liderar o mercado de Low-Code. O actual desafio é o de aumentar o ecosistema em 50 000 programadores até final de 2020.
+
+    img: "../../../images/gatsby-icon.png"
+    linkedin: https://www.linkedin.com/in/ricardonevesaraujo/
+    website: https://davidsopas.com
+    twitter: https://twitter.com/dsopas
+    
+
+  - name: Hélder Gonçalves
+    occupations:
+      - what: senior software engineer
+        where: outsystems
+    bio: É um Engenheiro de Software na OutSystems, que busca excelência no seu trabalho. Com uma mente curiosa e um espírito inquisitivo desafia os seus colegas a percepcionar se estão a tomar as decisões mais acertadas. Actualmente, desenvolve funcionalidades colaborativas que visam aumentar a produtividade e desempenho de equipas de trabalho OutSystems.
+
+    img: "../../../images/gatsby-icon.png"
+
+day: 2019-10-29
+start_time: 15h00
+end_time: 16h00
+place: B032
+type: 0-talk
+
+path: /software-engineering-at-outsystems.md
+---
+
+Todos os procuramos do que se faz bem lá fora esquecendo que em Portugal há equipas de classe mundial a inovar nos processos de engenharia de software. É isso que pretendemos dar a conhecer durante a nossa apresentação, quais os ingredientes essenciais para uma empresa de software nascida e baseada em Portugal
