+++ src/data/events/talks/test4.md
@@ -1,0 +1,27 @@
+---
+title: Breaking IoT
+
+speakers:
+  - name: David Sopas
+    occupations:
+      - what: Security Researcher
+        where: Checkmarx
+    bio: David Sopas é security researcher com mais de 10 anos de experiência. Atualmente lidera uma equipa de researchers da Checkmarx, é co-fundador da Char49 e Pentester da Cobalt. Foi reconhecido publicamente por encontrar falhas em empresas como Google, Microsoft, Adobe, eBay, Amazon e muitas outras. Alguns trabalhos de research do David já foram apresentados em conferências de segurança, como por exemplo, na Defcon, BSides e OWASP Appsec.
+
+    img: "../../../images/gatsby-icon.png"
+    linkedin: https://www.linkedin.com/in/miguelpoliveira/
+    website: https://davidsopas.com
+    twitter: https://twitter.com/dsopas
+
+day: 2019-10-29
+start_time: 18h10
+end_time: 18h40
+place: B032
+type: 0-talk
+
+path: /breaking-iot
+---
+
+A Internet da Coisas está a mudar os nossos hábitos de vida. Desde a utilização de wearables que nos informam que perdemos menos calorias que no dia anterior até aos dispositivos médicos que administram automaticamente a dose necessária para o doente. Esta "sede" de desenvolver dispositivos "inteligentes" está a gerar oportunidades aos
+utilizadores maliciosos de comprometerem muito mais que páginas web mas também o nosso quotidiano.
+Nesta palestra, David Sopas vai falar um pouco dos perigos associados ao IoT - desde a privacidade dos utilizadores até à segurança de dispositivos.
